refactoring and caching
